requirements.txt
Django>=2.2.3,<2.3.0
djangorestframework>=3.10.1,<3.11.0
psycopg2>=2.8.3,<2.9.0
django-extra-fields>=1.2.3,<1.3.0

flake8>=3.7.8,<3.8.0
